backend/requirements.txt
asgiref==3.8.1
bcrypt==4.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
pillow==11.2.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
